src/dockerProject/requirements.txt
#Werkzeug==2.0.3
#Flask-Cors==3.0.10
#requests==2.26.0
#pandas==1.3.4
#numpy==1.21.2
sqlalchemy
psycopg2
faker
kafka-python
confluent_kafka
paho-mqtt